Refactor : 프로젝트 모듈화
