+++ api/build.gradle
@@ -16,15 +16,6 @@ dependencies {
     implementation project(':domain')
     implementation project(':core')
 
-    // Spring Security
-    implementation 'org.springframework.boot:spring-boot-starter-security'
-    testImplementation 'org.springframework.security:spring-security-test'
-
-    // Jwt
-    implementation 'io.jsonwebtoken:jjwt-api:0.12.5'
-    implementation 'io.jsonwebtoken:jjwt-impl:0.12.5'
-    implementation 'io.jsonwebtoken:jjwt-jackson:0.12.5'
-
     // Swagger - SpringDoc
     implementation 'org.springdoc:springdoc-openapi-starter-webmvc-ui:2.3.0'
