+++ api/src/main/java/com/zero/pennywise/api/controller/TransactionController.java
@@ -3,7 +3,7 @@ package com.zero.pennywise.api.controller;
 import com.zero.pennywise.domain.model.request.transaction.TransactionDTO;
 import com.zero.pennywise.domain.model.request.transaction.TransactionInfoDTO;
 import com.zero.pennywise.domain.model.response.ResultResponse;
-import com.zero.pennywise.api.service.TransactionService;
+import com.zero.pennywise.core.service.TransactionService;
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.validation.Valid;
 import lombok.RequiredArgsConstructor;
