+++ core/src/main/java/com/zero/pennywise/core/config/BCryptEncoderConfig.java
@@ -1,4 +1,4 @@
-package com.zero.pennywise.api.config;
+package com.zero.pennywise.core.config;
 
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
