+++ core/src/main/java/com/zero/pennywise/core/service/AnalyzeService.java
