+++ core/src/main/java/com/zero/pennywise/core/service/NotificationService.java
@@ -1,4 +1,4 @@
-package com.zero.pennywise.api.service;
+package com.zero.pennywise.core.service;
 
 import com.zero.pennywise.domain.model.response.ResultResponse;
 import jakarta.servlet.http.HttpServletRequest;
