+++ core/src/main/java/com/zero/pennywise/core/service/impl/CategoryServiceImpl.java
@@ -1,11 +1,11 @@
-package com.zero.pennywise.api.service.impl;
+package com.zero.pennywise.core.service.impl;
 
-import com.zero.pennywise.api.component.FinanceFacade;
-import com.zero.pennywise.api.component.UserFacade;
+import com.zero.pennywise.core.service.CategoryService;
+import com.zero.pennywise.core.component.FinanceFacade;
+import com.zero.pennywise.core.component.UserFacade;
 import com.zero.pennywise.domain.entity.UserEntity;
 import com.zero.pennywise.domain.model.response.ResultResponse;
 import com.zero.pennywise.domain.model.type.SuccessResultCode;
-import com.zero.pennywise.api.service.CategoryService;
 import jakarta.servlet.http.HttpServletRequest;
 import java.util.List;
 import lombok.RequiredArgsConstructor;
